move sideTabs into /layout folder
