+++ static/app/views/replays/detail/layout/sideTabs.tsx
